requirements.txt
fastapi
uvicorn[standard]
pydantic
requests
python-dotenv
beautifulsoup4
scikit-learn
pandas
numpy